Initial stab at virtualbox debian ansible provisioning
Refactor Vagrant provisioning into subdirectories.

Testing the currently existing provisioning of ubuntu/trusty64 process
reveals that it hasn't worked since the move of the majority of the
process to blimp-engineroom.

The provisioing process contained in
<file:vagrant/debian/jessie64/Vagrantfile> will be the supported
configuration as identical to what we run the on Raspberry Pi2.

TO BE ADDRESSED
---------------
The vagrant ansible provisioning uses "sudo", whereas the RPi2 code
path uses "root" which needs to be aligned.

The vagrant ansible provisioning does not create "blimp_vars.sh", as I
am not sure of the right sequence: We want the result of ansible to be
a "raw" blimp without identity, yet we rely on "blimp_vars.sh" being
present at the initial upgrade time.